# ajk-mart-super-app/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AJKMart Super App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [mart, food]
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (vendor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /rides/{id}:
    get:
      operationId: getRide
      tags: [rides]
      summary: Get ride status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ride details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        category:
          type: string
        type:
          type: string
          enum: [mart, food]
        image:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        rating:
          type: number
        reviewCount:
          type: number
        inStock:
          type: boolean
        unit:
          type: string
        deliveryTime:
          type: string
      required: [id, name, price, category, type, inStock]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: number
      required: [products, total]
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        type:
          type: string
          enum: [mart, food]
        image:
          type: string
        vendorId:
          type: string
        unit:
          type: string
      required: [name, price, category, type, vendorId]
...
    Ride:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [car, bike]
        status:
          type: string
          enum: [searching, confirmed, in_progress, completed, cancelled]
        pickupAddress:
          type: string
        dropAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        dropLat:
          type: number
        dropLng:
          type: number
        fare:
          type: number
        distance:
          type: number
        riderId:
          type: string
        riderName:
          type: string
        riderPhone:
          type: string
        paymentMethod:
          type: string
          enum: [cash, wallet]
        createdAt:
          type: string
      required: [id, userId, type, status, fare, distance, paymentMethod, createdAt]
...